+++ fabtools/arch.py
@@ -13,7 +13,8 @@ from fabtools.utils import run_as_root
 
 
 def pkg_manager():
-    with settings(hide('running', 'stdout', 'stderr', 'warnings'), warn_only=True):
+    with settings(
+            hide('running', 'stdout', 'stderr', 'warnings'), warn_only=True):
         output = run('which yaourt', warn_only=True)
         if output.succeeded:
             manager = 'yaourt'
@@ -30,7 +31,9 @@ def update_index(quiet=True):
 
     manager = pkg_manager()
     if quiet:
-        with settings(hide('running', 'stdout', 'stderr', 'warnings'), warn_only=True):
+        with settings(
+                hide('running', 'stdout', 'stderr', 'warnings'),
+                warn_only=True):
             run_as_root("%(manager)s -Sy" % locals())
     else:
         run_as_root("%(manager)s -Sy" % locals())
@@ -49,7 +52,8 @@ def is_installed(pkg_name):
     Check if an Arch Linux package is installed.
     """
 
-    with settings(hide('running', 'stdout', 'stderr', 'warnings'), warn_only=True):
+    with settings(
+            hide('running', 'stdout', 'stderr', 'warnings'), warn_only=True):
         res = run("pacman -Q %(pkg_name)s" % locals())
         return res.succeeded