+++ fabtools/files.py
@@ -213,7 +213,7 @@ class watch(object):
         from fabric.contrib.files import comment, uncomment
 
         from fabtools.files import watch
-        from fabtools.services import restart
+        from fabtools.service import restart
 
         # Edit configuration file
         with watch('/etc/daemon.conf') as config:
@@ -231,7 +231,7 @@ class watch(object):
         from fabric.contrib.files import comment, uncomment
 
         from fabtools.files import watch
-        from fabtools.services import restart
+        from fabtools.service import restart
 
         with watch('/etc/daemon.conf', callback=partial(restart, 'daemon')):
             uncomment('/etc/daemon.conf', 'someoption')
@@ -295,7 +295,8 @@ def copy(source, destination, recursive=False, use_sudo=False):
     """
     func = use_sudo and run_as_root or run
     options = '-r ' if recursive else ''
-    func('/bin/cp {0}{1} {2}'.format(options, quote(source), quote(destination)))
+    func('/bin/cp {0}{1} {2}'.format(
+        options, quote(source), quote(destination)))
 
 
 def move(source, destination, use_sudo=False):