+++ fabtools/mysql.py
@@ -16,7 +16,9 @@ from fabtools.utils import run_as_root
 
 
 def query(query, use_sudo=True, **kwargs):
-
+    """
+    Run a MySQL query.
+    """
     family = distrib_family()
     if family == 'debian':
         from fabtools.deb import install, is_installed
@@ -30,6 +32,7 @@ def query(query, use_sudo=True, **kwargs):
     user = kwargs.get('mysql_user') or env.get('mysql_user')
     password = kwargs.get('mysql_password') or env.get('mysql_password')
     func_mysql = 'mysql'
+    mysql_host = kwargs.get('mysql_host') or env.get('mysql_host')
 
     options = [
         '--batch',
@@ -43,6 +46,9 @@ def query(query, use_sudo=True, **kwargs):
             install('sshpass')
         func_mysql = 'sshpass -p %(password)s mysql' % {'password': password}
         options.append('--password')
+        options.append('--password=%s' % quote(password))
+    if mysql_host:
+        options.append('--host=%s' % quote(mysql_host))
     options = ' '.join(options)
 
     return func('%(cmd)s %(options)s --execute=%(query)s' % {
@@ -56,15 +62,13 @@ def user_exists(name, host='localhost', **kwargs):
     """
     Check if a MySQL user exists.
     """
-    with settings(hide('running', 'stdout', 'stderr', 'warnings'), warn_only=True):
+    with settings(
+            hide('running', 'stdout', 'stderr', 'warnings'), warn_only=True):
         res = query("""
             use mysql;
             SELECT COUNT(*) FROM user
                 WHERE User = '%(name)s' AND Host = '%(host)s';
-            """ % {
-                'name': name,
-                'host': host,
-            }, **kwargs)
+            """ % {'name': name, 'host': host}, **kwargs)
     return res.succeeded and (int(res) == 1)
 
 
@@ -82,11 +86,13 @@ def create_user(name, password, host='localhost', **kwargs):
 
     """
     with settings(hide('running')):
-        query("CREATE USER '%(name)s'@'%(host)s' IDENTIFIED BY '%(password)s';" % {
-            'name': name,
-            'password': password,
-            'host': host
-        }, **kwargs)
+        query(
+            "CREATE USER '%(name)s'@'%(host)s' IDENTIFIED BY '%(password)s';" %
+            {
+                'name': name,
+                'password': password,
+                'host': host
+            }, **kwargs)
     puts("Created MySQL user '%s'." % name)
 
 
@@ -94,7 +100,8 @@ def database_exists(name, **kwargs):
     """
     Check if a MySQL database exists.
     """
-    with settings(hide('running', 'stdout', 'stderr', 'warnings'), warn_only=True):
+    with settings(
+            hide('running', 'stdout', 'stderr', 'warnings'), warn_only=True):
         res = query("SHOW DATABASES LIKE '%(name)s';" % {
             'name': name
         }, **kwargs)