+++ fabtools/require/portage.py
@@ -14,7 +14,6 @@ from fabtools.portage import (
     install,
     is_installed,
     uninstall,
-    update_index,
 )