+++ fabtools/require/python.py
@@ -40,21 +40,7 @@ def setuptools(version=MIN_SETUPTOOLS_VERSION, python_cmd='python'):
     .. _setuptools: http://pythonhosted.org/setuptools/
     """
 
-    from fabtools.require.deb import package as require_deb_package
-    from fabtools.require.rpm import package as require_rpm_package
-
     if not is_setuptools_installed(python_cmd=python_cmd):
-        family = distrib_family()
-
-        if family == 'debian':
-            require_deb_package('python-dev')
-        elif family == 'redhat':
-            require_rpm_package('python-devel')
-        elif family == 'arch':
-            pass  # ArchLinux installs header with base package
-        else:
-            raise UnsupportedFamily(supported=['debian', 'redhat', 'arch'])
-
         install_setuptools(python_cmd=python_cmd)
 
 
@@ -68,7 +54,7 @@ def pip(version=MIN_PIP_VERSION, pip_cmd='pip', python_cmd='python'):
     .. _pip: http://www.pip-installer.org/
     """
     setuptools(python_cmd=python_cmd)
-    if not is_pip_installed(version, pip_cmd=pip_cmd):
+    if not is_pip_installed(version, python_cmd=python_cmd, pip_cmd=pip_cmd):
         install_pip(python_cmd=python_cmd)
 
 
@@ -102,8 +88,9 @@ def package(pkg_name, url=None, pip_cmd='pip', python_cmd='python',
     .. _pip installer: http://www.pip-installer.org/
     """
     pip(MIN_PIP_VERSION, python_cmd=python_cmd)
-    if not is_installed(pkg_name, pip_cmd=pip_cmd):
+    if not is_installed(pkg_name, python_cmd=python_cmd, pip_cmd=pip_cmd):
         install(url or pkg_name,
+                python_cmd=python_cmd,
                 pip_cmd=pip_cmd,
                 allow_external=[url or pkg_name] if allow_external else [],
                 allow_unverified=[url or pkg_name] if allow_unverified else [],
@@ -131,9 +118,11 @@ def packages(pkg_list, pip_cmd='pip', python_cmd='python',
 
     pip(MIN_PIP_VERSION, python_cmd=python_cmd)
 
-    pkg_list = [pkg for pkg in pkg_list if not is_installed(pkg, pip_cmd=pip_cmd)]
+    pkg_list = [
+        pkg for pkg in pkg_list if not is_installed(pkg, python_cmd=python_cmd, pip_cmd=pip_cmd)]
     if pkg_list:
         install(pkg_list,
+                python_cmd=python_cmd,
                 pip_cmd=pip_cmd,
                 allow_external=allow_external,
                 allow_unverified=allow_unverified,
@@ -163,13 +152,15 @@ def requirements(filename, pip_cmd='pip', python_cmd='python',
     .. _requirements file: http://www.pip-installer.org/en/latest/requirements.html
     """
     pip(MIN_PIP_VERSION, python_cmd=python_cmd)
-    install_requirements(filename, pip_cmd=pip_cmd, allow_external=allow_external,
+    install_requirements(filename, python_cmd=python_cmd, pip_cmd=pip_cmd,
+                         allow_external=allow_external,
                          allow_unverified=allow_unverified, **kwargs)
 
 
 def virtualenv(directory, system_site_packages=False, venv_python=None,
                use_sudo=False, user=None, clear=False, prompt=None,
-               virtualenv_cmd='virtualenv', pip_cmd='pip', python_cmd='python'):
+               virtualenv_cmd='virtualenv', pip_cmd='pip',
+               python_cmd='python'):
     """
     Require a Python `virtual environment`_.
 
@@ -182,7 +173,8 @@ def virtualenv(directory, system_site_packages=False, venv_python=None,
     .. _virtual environment: http://www.virtualenv.org/
     """
 
-    package('virtualenv', use_sudo=True, pip_cmd=pip_cmd, python_cmd=python_cmd)
+    package('virtualenv', use_sudo=True, pip_cmd=pip_cmd,
+            python_cmd=python_cmd)
 
     if not virtualenv_exists(directory):
         create_virtualenv(